+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Added the diffy project:update command.
+* Add possibility to create and update multiple projects by a JSON array.
 
 ### 0.1.0 - 2020/Jan/10